# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Learning Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  instructor: 'instructor',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  replitId: string;
  name: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  role: UserRole;
  createdAt: string;
}

export type UserRoleUpdateRole = typeof UserRoleUpdateRole[keyof typeof UserRoleUpdateRole];


export const UserRoleUpdateRole = {
  student: 'student',
  instructor: 'instructor',
  admin: 'admin',
} as const;

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}

export interface Course {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  instructorId: string;
  /** @nullable */
  instructorName?: string | null;
  /** @nullable */
  price?: number | null;
  isFree: boolean;
  lessonCount?: number;
  enrollmentCount?: number;
  createdAt: string;
}

export interface CourseInput {
  /** @minLength 1 */
  title: string;
  description: string;
  thumbnailUrl?: string;
  price?: number;
  isFree?: boolean;
}

export interface CourseUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  thumbnailUrl?: string;
  price?: number;
  isFree?: boolean;
}

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  videoUrl: string;
  /** @nullable */
  description?: string | null;
  orderIndex: number;
  completed?: boolean;
  createdAt: string;
}

export interface LessonInput {
  /** @minLength 1 */
  title: string;
  videoUrl: string;
  description?: string;
  orderIndex?: number;
}

export interface LessonProgress {
  lessonId: number;
  completed: boolean;
}

export interface CourseProgress {
  courseId: number;
  totalLessons: number;
  completedLessons: number;
  progressPercent: number;
}

export interface Enrollment {
  id: number;
  userId: string;
  courseId: number;
  /** @nullable */
  courseTitle?: string | null;
  /** @nullable */
  courseThumbnailUrl?: string | null;
  progressPercent?: number;
  enrolledAt: string;
}

export interface EnrollmentInput {
  courseId: number;
}

export interface Quiz {
  id: number;
  courseId: number;
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  createdAt: string;
}

export type QuizInputCorrectAnswer = typeof QuizInputCorrectAnswer[keyof typeof QuizInputCorrectAnswer];


export const QuizInputCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface QuizInput {
  /** @minLength 1 */
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: QuizInputCorrectAnswer;
}

export type QuizSubmissionAnswersItemAnswer = typeof QuizSubmissionAnswersItemAnswer[keyof typeof QuizSubmissionAnswersItemAnswer];


export const QuizSubmissionAnswersItemAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export type QuizSubmissionAnswersItem = {
  quizId: number;
  answer: QuizSubmissionAnswersItemAnswer;
};

export interface QuizSubmission {
  answers: QuizSubmissionAnswersItem[];
}

export interface QuizResult {
  score: number;
  total: number;
  percentage: number;
}

export interface Assignment {
  id: number;
  courseId: number;
  title: string;
  description: string;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export interface AssignmentInput {
  /** @minLength 1 */
  title: string;
  description: string;
  dueDate?: string;
}

export interface Submission {
  id: number;
  assignmentId: number;
  userId: string;
  /** @nullable */
  userName?: string | null;
  answer: string;
  /** @nullable */
  fileUrl?: string | null;
  submittedAt: string;
}

export interface SubmissionInput {
  answer: string;
  fileUrl?: string;
}

export interface DiscussionPost {
  id: number;
  courseId: number;
  userId: string;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userAvatarUrl?: string | null;
  message: string;
  /** @nullable */
  parentId?: number | null;
  replies?: DiscussionPost[];
  createdAt: string;
}

export interface DiscussionInput {
  /** @minLength 1 */
  message: string;
  parentId?: number;
}

export interface DashboardSummary {
  role: string;
  totalCourses?: number;
  totalEnrollments?: number;
  totalUsers?: number;
  totalStudents?: number;
  totalInstructors?: number;
  myCoursesCount?: number;
  enrolledCoursesCount?: number;
  completedCoursesCount?: number;
  averageProgress?: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  enrollment: 'enrollment',
  completion: 'completion',
  submission: 'submission',
  discussion: 'discussion',
  course_created: 'course_created',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  message: string;
  /** @nullable */
  courseTitle?: string | null;
  createdAt: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListCoursesParams = {
instructorId?: string;
};