# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Learning Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [Courses]
      summary: List all courses
      parameters:
        - name: instructorId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [Courses]
      summary: Create a new course (instructor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [Courses]
      summary: Get a course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
    patch:
      operationId: updateCourse
      tags: [Courses]
      summary: Update a course (instructor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [Courses]
      summary: Delete a course (instructor or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Lessons ──────────────────────────────────────────────────────────────────
...
  /courses/{courseId}/assignments:
    get:
      operationId: listAssignments
      tags: [Assignments]
      summary: List assignments for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [Assignments]
      summary: Create an assignment (instructor only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}/submissions:
    get:
      operationId: listSubmissions
      tags: [Assignments]
      summary: List submissions for an assignment (instructor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: submitAssignment
      tags: [Assignments]
      summary: Submit an assignment (student)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # ── Discussions ────────────────────────────────────────────────────────────────
  /courses/{courseId}/discussions:
    get:
      operationId: listDiscussions
      tags: [Discussions]
      summary: List discussion posts for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of discussion posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscussionPost"
    post:
      operationId: createDiscussion
      tags: [Discussions]
      summary: Post a message to a course discussion
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscussionPost"
  /discussions/{id}:
    delete:
      operationId: deleteDiscussion
      tags: [Discussions]
      summary: Delete a discussion post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard ─────────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Course:
      type: object
      required: [id, title, description, instructorId, isFree, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: [string, 'null']
        instructorId:
          type: string
        instructorName:
          type: [string, 'null']
        price:
          type: [number, 'null']
        isFree:
          type: boolean
        lessonCount:
          type: integer
        enrollmentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CourseInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        thumbnailUrl:
          type: string
        price:
          type: number
        isFree:
          type: boolean
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        thumbnailUrl:
          type: string
        price:
          type: number
        isFree:
          type: boolean
...
    Assignment:
      type: object
      required: [id, courseId, title, description, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        dueDate:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    AssignmentInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        dueDate:
          type: string
    Submission:
      type: object
      required: [id, assignmentId, userId, answer, submittedAt]
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        userId:
          type: string
        userName:
          type: [string, 'null']
        answer:
          type: string
        fileUrl:
          type: [string, 'null']
        submittedAt:
          type: string
          format: date-time
    SubmissionInput:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
        fileUrl:
          type: string
    DiscussionPost:
      type: object
      required: [id, courseId, userId, message, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        userId:
          type: string
        userName:
          type: [string, 'null']
        userAvatarUrl:
          type: [string, 'null']
        message:
          type: string
        parentId:
          type: [integer, 'null']
        replies:
          type: array
          items:
            $ref: "#/components/schemas/DiscussionPost"
        createdAt:
          type: string
          format: date-time
    DiscussionInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        parentId:
          type: integer
...