# lib/api-zod/src/generated/types/courseUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Learning Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CourseUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  thumbnailUrl?: string;
  price?: number;
  isFree?: boolean;
}