# lib/api-zod/src/generated/types/discussionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Learning Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface DiscussionInput {
  /** @minLength 1 */
  message: string;
  parentId?: number;
}